Modify ip_broadcast and ip_network
This changes to behavior of these two functions so that they return a value
which continues to carry the prefix length.  To turn an offset address into a
simple IP network address, without a prefix, the new way is to chain the functions:

ip_address(ip_network("10.10.10.10/19")) == "10.10.8.0"

Likewise for ip_broadcast.